Moved errors out into dedicated fle
